requirements.txt
opencv-python
ultralytics
supervision
numpy
torchreid-0.2.5
gdown
torch>=2.5.0
torchvision

# LLM support
openai
langchain
langchain-community
langchain-core
tiktoken

# Optional (faster LLM)
groq
Flask
python-dotenv
psutil
PyYAML
requests
scipy
pydantic
pydantic-settings
pydantic-core
anyio
httpx
sniffio
idna
certifi
h11
httpcore
distro
blinker
click
itsdangerous
Jinja2
MarkupSafe
Werkzeug